+++ ecs/resource.go
@@ -8,6 +8,14 @@ type Resource[T any] struct {
 	id    ResID
 }
 
+// New creates a new [Resource]. It is safe to call on an uninitialized instance.
+// It is a helper method, intended to avoid repeated listing of type parameters.
+//
+// Due to value receiver, the return type must be assigned back to the original value.
+func (Resource[T]) New(world *World) Resource[T] {
+	return NewResource[T](world)
+}
+
 // NewResource creates a new [Resource] mapper for a resource type.
 // This does not add a resource to the world, but only creates a mapper for resource access!
 //