+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Adds custom event support for the event system (#340, #348)
 - Adds method `World.Shrink` for freeing memory that exceeds current requirements (#323)
 - World lock and filters are concurrency-safe, allowing for concurrent query execution (#360)
+- Adds method `Resource[T].New` as a shortcut to avoid repeated parameter specification (#393)
 
 ### Performance